Push select gateways to sub module
